# parse_styles optimization: use string_id for CSS property names