--- a/include/litehtml/el_text.h
+++ b/include/litehtml/el_text.h
@@ -17,7 +17,7 @@ namespace litehtml
 		el_text(const char* text, const std::shared_ptr<litehtml::document>& doc);
 
 		void				get_text(string& text) override;
-		const char*			get_style_property(const char* name, bool inherited, const char* def = nullptr) const override;
+		const char*			get_style_property(string_id name, bool inherited, const char* def = nullptr) const override;
 		void				parse_styles(bool is_reparse) override;
         bool				is_text() const override { return true; }