--- a/src/document.cpp
+++ b/src/document.cpp
@@ -759,11 +759,9 @@ void litehtml::document::create_node(void* gnode, elements_vector& elements, boo
 		break;
 	case GUMBO_NODE_TEXT:
 		{
-			std::wstring str;
-			std::wstring str_in = (const wchar_t*) (utf8_to_wchar(node->v.text.text));
 			if (!parseTextNode)
 			{
-				elements.push_back(std::make_shared<el_text>(litehtml_from_wchar(str_in).c_str(), shared_from_this()));
+				elements.push_back(std::make_shared<el_text>(node->v.text.text, shared_from_this()));
 			}
 			else
 			{