--- a/src/el_div.cpp
+++ b/src/el_div.cpp
@@ -12,7 +12,7 @@ void litehtml::el_div::parse_attributes()
 	const char* str = get_attr("align");
 	if(str)
 	{
-		m_style.add_property("text-align", str, 0, false, this);
+		m_style.add_property(_text_align_, str, 0, false, this);
 	}
 	html_tag::parse_attributes();
 }