--- a/src/el_td.cpp
+++ b/src/el_td.cpp
@@ -12,7 +12,7 @@ void litehtml::el_td::parse_attributes()
 	const char* str = get_attr("width");
 	if(str)
 	{
-		m_style.add_property("width", str, nullptr, false, this);
+		m_style.add_property(_width_, str, nullptr, false, this);
 	}
 	str = get_attr("background");
 	if(str)
@@ -20,24 +20,24 @@ void litehtml::el_td::parse_attributes()
 		string url = "url('";
 		url += str;
 		url += "')";
-		m_style.add_property("background-image", url.c_str(), nullptr, false, this);
+		m_style.add_property(_background_image_, url.c_str(), nullptr, false, this);
 	}
 	str = get_attr("align");
 	if(str)
 	{
-		m_style.add_property("text-align", str, nullptr, false, this);
+		m_style.add_property(_text_align_, str, nullptr, false, this);
 	}
 
 	str = get_attr("bgcolor");
 	if (str)
 	{
-		m_style.add_property("background-color", str, nullptr, false, this);
+		m_style.add_property(_background_color_, str, nullptr, false, this);
 	}
 
 	str = get_attr("valign");
 	if(str)
 	{
-		m_style.add_property("vertical-align", str, nullptr, false, this);
+		m_style.add_property(_vertical_align_, str, nullptr, false, this);
 	}
 	html_tag::parse_attributes();
 }