--- a/src/el_tr.cpp
+++ b/src/el_tr.cpp
@@ -12,17 +12,17 @@ void litehtml::el_tr::parse_attributes()
 	const char* str = get_attr("align");
 	if(str)
 	{
-		m_style.add_property("text-align", str, nullptr, false, this);
+		m_style.add_property(_text_align_, str, nullptr, false, this);
 	}
 	str = get_attr("valign");
 	if(str)
 	{
-		m_style.add_property("vertical-align", str, nullptr, false, this);
+		m_style.add_property(_vertical_align_, str, nullptr, false, this);
 	}
 	str = get_attr("bgcolor");
 	if (str)
 	{
-		m_style.add_property("background-color", str, nullptr, false, this);
+		m_style.add_property(_background_color_, str, nullptr, false, this);
 	}
 	html_tag::parse_attributes();
 }