## Before this PR

I am trying to improve rendering performance on Obama wiki.
I found out that the main culprit is document::createFromString, and within it apply_stylesheet and parse_styles.
The main problem with parse_styles turned out to be cairo_container::text_width.
So I switched to win32_container, which slashed about 3.7 sec from parse_styles:
text_width aggregate time on Obama wiki:
cairo_container: ~ 3.8 sec
win32_container: 111 ms

## This PR

This PR further improves performance of parse_styles, even though the main problem now is apply_stylesheet:
Before:
```
> apply_stylesheet: 3000 ms
> parse_styles: 750 ms
document::createFromString: 4172 ms
```
After:
```
> apply_stylesheet: 2969 ms      (same)
> parse_styles: 468 ms     <---  improvement
document::createFromString: 3906 ms
```

To achieve this I use integers (`string_id`) instead of strings for CSS property names.

`string_id` is an open enum. Known strings like "border" have predefined names (`_border_`) and numeric values.
Unknown strings (like class names) don't have names, only autogenerated numbers (the number is generated by _id function).
Currently I use string_id only for property names, but I plan to also use them for other strings like property values, tag names, attributes, class names and element ids.
There are two functions:
```
_id: string -> string_id
_s:  string_id -> string
```

**CAUTION**: After this change litehtml cannot be used before main because of `init`, see string_id.cpp.

Collateral changes: removed parse_short_border because it duplicated already existing code and was never called.

## After this PR

This PR is relatively benign, but I think more drastic changes are required to further improve styling performance:
* using string_id for CSS property values will result in flattening of various enums in types.h: instead of style_display, list_style_type, etc there will be only string_id
* parse_stylesheet and parse_styles might need to be merged; this is also required to correctly handle situations like `p{width:100px;width:foo}` (with current approach `width:100px` is lost)
* css selection code might need to be rewritten in a non-recursive way

Are you OK with such big changes if they gain performance?