@tordex 2022-11-08
Great work! Thank you!
> Are you OK with such big changes if they gain performance?

Yes, sure. Performance is very important.