@stasoid 2022-11-08
> Yes, sure. Performance is very important.

Cool, thank you.