@cj-mcdonald 2022-11-08
Hi, the new string_id map looks like a good candidate for std:: unordered_map to give better performance, with lots of direct keyed access, no deletions and no ordered iterations. May be worth trying.